+++ shellSort.h
@@ -1,16 +1,16 @@
 #pragma once
 #include <vector>
 
-void shellSort(std::vector<int>& arr, int minNumber, int maxNumber) {
+void shellSort(std::vector<long long>& arr, long long minNumber, long long maxNumber) {
     if (arr.empty()) return;
 
     if (maxNumber - minNumber >= 100000) return;
     
-    int n = arr.size();
-    for (int gap = n/2; gap > 0; gap /= 2) {
-        for (int i = gap; i < n; ++i) {
-            int temp = arr[i];
-            int j;
+    long long n = arr.size();
+    for (long long gap = n/2; gap > 0; gap /= 2) {
+        for (long long i = gap; i < n; ++i) {
+            long long temp = arr[i];
+            long long j;
             for (j = i; j >= gap && arr[j - gap] > temp; j -= gap) {
                 arr[j] = arr[j - gap];
             }
